pyproject.toml
[project]
name = "omni2"
version = "0.1.0"
description = "OMNI2 Bridge - Intelligent MCP orchestration layer with LLM-based routing"
authors = [
    { name = "Avi Cohen", email = "avicoiot@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["mcp", "model-context-protocol", "llm", "orchestration", "slack-bot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Async HTTP Client
    "httpx>=0.25.0",
    
    # MCP Integration
    "fastmcp>=0.2.0",            # FastMCP for MCP protocol communication
    
    # Database
    "asyncpg>=0.29.0",          # PostgreSQL async driver
    "sqlalchemy>=2.0.23",        # ORM
    "alembic>=1.12.1",           # Migrations (future)
    
    # LLM Integration
    "anthropic>=0.7.0",          # Claude API
    
    # Slack Integration
    "slack-sdk>=3.26.0",         # Slack API
    "slack-bolt>=1.18.0",        # Slack Bolt framework
    
    # Configuration
    "pydantic>=2.5.0",           # Data validation
    "pydantic-settings>=2.1.0",  # Settings management
    "pyyaml>=6.0.1",             # YAML parsing
    
    # Utilities
    "python-dotenv>=1.0.0",      # Environment variables
    "orjson>=3.9.10",            # Fast JSON
    "python-multipart>=0.0.6",   # File uploads
    "psutil>=5.9.0",             # System metrics
    
    # Logging & Monitoring
    "structlog>=23.2.0",         # Structured logging
    "prometheus-client>=0.19.0", # Metrics (Phase 2)
    
    # Security
    "python-jose[cryptography]>=3.3.0",  # JWT tokens (Phase 2)
    "passlib[bcrypt]>=1.7.4",            # Password hashing (Phase 2)
    
    # Date/Time
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",             # For testing httpx clients
    
    # Code Quality
    "ruff>=0.1.6",               # Linting & formatting
    "mypy>=1.7.1",               # Type checking
    "black>=23.12.0",            # Code formatter (if not using ruff)
    "isort>=5.13.0",             # Import sorting
    
    # Development Tools
    "ipython>=8.18.1",           # Better REPL
    "ipdb>=0.13.13",             # Debugger
    "watchfiles>=0.21.0",        # File watcher for auto-reload
    
    # Documentation
    "mkdocs>=1.5.3",             # Documentation generator
    "mkdocs-material>=9.5.0",    # Material theme
]

production = [
    # Production-only dependencies
    "gunicorn>=21.2.0",          # Production WSGI server
    "setproctitle>=1.3.3",       # Process naming
]

[project.urls]
Homepage = "https://github.com/aviciot/OmniV2"
Repository = "https://github.com/aviciot/OmniV2.git"
Issues = "https://github.com/aviciot/OmniV2/issues"
Documentation = "https://github.com/aviciot/OmniV2/wiki"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ============================================================
# Ruff Configuration (Linting & Formatting)
# ============================================================
[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

# Enable specific rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value used in comparison
    "TCH001",  # Move imports into TYPE_CHECKING block
    "TCH002",  # Move imports into TYPE_CHECKING block
    "TCH003",  # Move imports into TYPE_CHECKING block
]

# Exclude patterns
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
    "migrations",  # Alembic migrations
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["app"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# ============================================================
# MyPy Configuration (Type Checking)
# ============================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

# Per-module options
[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "slack_sdk.*",
    "slack_bolt.*",
    "asyncpg.*",
]
ignore_missing_imports = true

# ============================================================
# Pytest Configuration (Testing)
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Coverage options
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================
# Coverage Configuration
# ============================================================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# ============================================================
# Black Configuration (if not using Ruff formatter)
# ============================================================
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
  | migrations
)/
'''

# ============================================================
# isort Configuration (if not using Ruff)
# ============================================================
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
